feat(mindmap): fix learningMode for roadmap, UI for completed node, remove redundant mindmap.utils.ts
